Renamed Patient Identifier

+++ src/main/java/com/alienworkspace/cdr/model/dto/patient/PatientProgramDto.java
@@ -57,7 +57,7 @@ public class PatientProgramDto {
             description = "Outcome Concept Id",
             example = "1"
     )
-    private Long outcomeConceptId;
+    private Integer outcomeConceptId;
 
     @Schema(
             description = "Outcome Comment",
